+++ kinto_http/tests/test_client.py
@@ -510,6 +510,40 @@ class RecordTest(unittest.TestCase):
             session=self.session, bucket='mybucket',
             collection='mycollection')
 
+    def test_create_record_refresh_records_timestamp(self):
+        mock_response(self.session)
+        self.client._records_timestamp = mock.sentinel.records_timestamp
+        self.client.create_record({'foo': 'bar'})
+        assert self.client._records_timestamp == {}
+
+    def test_update_record_refresh_records_timestamp(self):
+        mock_response(self.session)
+        self.client._records_timestamp = mock.sentinel.records_timestamp
+        self.client.update_record(data={'id': '1234'},
+                                  collection='good_collection',
+                                  permissions=mock.sentinel.permissions)
+        assert self.client._records_timestamp == {}
+
+    def test_patch_record_refresh_records_timestamp(self):
+        mock_response(self.session)
+        self.client._records_timestamp = mock.sentinel.records_timestamp
+        self.client.patch_record(data={'id': '1234'},
+                                 collection='good_collection',
+                                 permissions=mock.sentinel.permissions)
+        assert self.client._records_timestamp == {}
+
+    def test_delete_record_refresh_records_timestamp(self):
+        mock_response(self.session)
+        self.client._records_timestamp = mock.sentinel.records_timestamp
+        self.client.delete_record('1234')
+        assert self.client._records_timestamp == {}
+
+    def test_delete_records_refresh_records_timestamp(self):
+        mock_response(self.session)
+        self.client._records_timestamp = mock.sentinel.records_timestamp
+        self.client.delete_records('1234')
+        assert self.client._records_timestamp == {}
+
     def test_record_id_is_given_after_creation(self):
         mock_response(self.session, data={'id': 5678})
         record = self.client.create_record({'foo': 'bar'})
@@ -644,6 +678,20 @@ class RecordTest(unittest.TestCase):
         assert timestamp == '12345'
 
         timestamp = self.client.get_records_timestamp("bar")
+        assert timestamp == '67890'
+
+    def test_records_timestamp_cache_can_be_refreshed(self):
+        mock_response(self.session, headers={"ETag": '"12345"'})
+
+        timestamp = self.client.get_records_timestamp("foo")
+        assert timestamp == '12345'
+
+        mock_response(self.session, headers={"ETag": '"67890"'})
+        timestamp = self.client.get_records_timestamp("foo")
+        assert timestamp == '12345'
+
+        self.client.refresh()
+        timestamp = self.client.get_records_timestamp("foo")
         assert timestamp == '67890'
 
     def test_pagination_is_followed(self):